Define separate interface AmqpConsumerClientInterface
The change address two problem:

- We have a lot of services which are using messagestores while they only have consumers.
 The problem with this, we make almost 6 Million request per day which are totally useless: https://bit.ly/2T5LhY0
- Make it harder to forget define messagestore for the client if we are start using publishers in a service
